prevent duplicate slugs when creating new post
When creating a post with a filename identfical to a previous post (even
if the date is different), load the original post for editing instead.

If you want to create a new post, rather than editing the existing one,
you'll need to provide a different filename on the command line.